Make a refactor of Login and Routes components.
