+++ react-app/src/app/screens/Login/index.js
@@ -5,17 +5,17 @@ import { Redirect } from 'react-router-dom';
 
 import ActionCreators from '../../../redux/Login/actions';
 
-import LoginForm from './components/LoginForm';
+import Form from './layout';
 
-class LoginFormContainer extends Component {
+class Login extends Component {
   submit = values => this.props.submit(values);
 
   render() {
-    return this.props.isLogged ? <Redirect to="/" /> : <LoginForm onSubmit={this.submit} />;
+    return this.props.isLogged ? <Redirect to="/" /> : <Form onSubmit={this.submit} />;
   }
 }
 
-LoginFormContainer.propTypes = {
+Login.propTypes = {
   submit: PropTypes.func.isRequired,
   isLogged: PropTypes.bool
 };
@@ -32,4 +32,4 @@ const mapDispatchToProps = dispatch => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(LoginFormContainer);
+)(Login);
