+++ react-app/src/utils/index.js
@@ -1,0 +1,3 @@
+export const getStorageItem = value => localStorage.getItem(value);
+
+export const setStorageItem = (key, value) => localStorage.setItem(key, value);
